refactor: move engine actor to engine.rs, make BrowserApp a thin GUI shell (#72)
- Move EngineCmd, EngineHandle, and run_engine_actor() from
  browser_daemon.rs into engine.rs so both main.rs and browser_daemon.rs
  can share the actor pattern
- Add EngineHandle::spawn() convenience constructor
- Refactor BrowserApp to hold engine::EngineHandle instead of raw
  BrowserEngine; remove all direct calls to fetch_and_process,
  process_html_with_cache, and reqwest in main.rs
- Add tick_promise and click_promise to BrowserApp, matching the
  DaemonBrowserApp pattern for bounded JS ticks and click handling
- Image loading now sends EngineCmd::LoadImage to the actor on
  promise resolution, same as DaemonBrowserApp
- Remove duplicate EngineCmd/EngineHandle definitions from
  browser_daemon.rs; update to import from engine module
