+++ src/engine.rs
@@ -913,6 +913,246 @@ fn hit_test(x: f32, y: f32, r: &layout::Rect) -> bool {
     x >= r.x && x <= r.x + r.width && y >= r.y && y <= r.y + r.height
 }
 
+// ── Engine actor ──────────────────────────────────────────────────────────────
+
+use std::sync::mpsc;
+
+/// Commands sent to the engine actor thread.
+/// All engine work (including blocking HTTP fetches and JS) happens on that thread.
+pub enum EngineCmd {
+    Navigate {
+        url: String,
+        width: f32,
+        reply: mpsc::Sender<Result<PageResult, String>>,
+    },
+    ReRender {
+        hovered_id: Option<String>,
+        focused_id: Option<String>,
+        width: f32,
+        reply: mpsc::Sender<Result<PageResult, String>>,
+    },
+    Click {
+        x: f32,
+        y: f32,
+        reply: mpsc::Sender<Vec<ClickResult>>,
+    },
+    TypeText {
+        text: String,
+    },
+    EvaluateJs {
+        script: String,
+        reply: mpsc::Sender<String>,
+    },
+    Screenshot {
+        reply: mpsc::Sender<Option<Vec<u8>>>,
+    },
+    DomTree {
+        reply: mpsc::Sender<String>,
+    },
+    LayoutTree {
+        reply: mpsc::Sender<String>,
+    },
+    ComputedStyle {
+        selector: String,
+        reply: mpsc::Sender<HashMap<String, String>>,
+    },
+    GetPage {
+        reply: mpsc::Sender<Option<ApiPageResponse>>,
+    },
+    GetElements {
+        reply: mpsc::Sender<Vec<ApiElement>>,
+    },
+    LoadImage {
+        url: String,
+        bytes: Vec<u8>,
+    },
+    Tick {
+        timestamp: f64,
+        deadline: Option<f64>,
+        reply: mpsc::Sender<bool>,
+    },
+    #[allow(dead_code)]
+    Shutdown,
+}
+
+/// Run the engine actor — owns `BrowserEngine` exclusively.
+/// Processes commands sequentially from the receiver.
+/// This guarantees all `reqwest::blocking` calls and `thread_local!` JS state
+/// are confined to a single thread.
+pub fn run_engine_actor(rx: mpsc::Receiver<EngineCmd>) {
+    let mut eng = BrowserEngine::new();
+    for cmd in rx {
+        match cmd {
+            EngineCmd::Navigate { url, width, reply } => {
+                let result = eng.navigate(&url, width);
+                let _ = reply.send(result);
+            }
+            EngineCmd::ReRender { hovered_id, focused_id, width, reply } => {
+                let result = eng.re_render(hovered_id.as_deref(), focused_id.as_deref(), width);
+                let _ = reply.send(result);
+            }
+            EngineCmd::Click { x, y, reply } => {
+                let result = eng.click(x, y);
+                let _ = reply.send(result);
+            }
+            EngineCmd::TypeText { text } => {
+                eng.type_text(&text);
+            }
+            EngineCmd::EvaluateJs { script, reply } => {
+                let result = eng.evaluate_js(&script);
+                let _ = reply.send(result);
+            }
+            EngineCmd::Screenshot { reply } => {
+                let png = eng.screenshot().and_then(|pm| pm.encode_png().ok());
+                let _ = reply.send(png);
+            }
+            EngineCmd::DomTree { reply } => {
+                let _ = reply.send(eng.dom_tree());
+            }
+            EngineCmd::LayoutTree { reply } => {
+                let _ = reply.send(eng.layout_tree());
+            }
+            EngineCmd::ComputedStyle { selector, reply } => {
+                let _ = reply.send(eng.computed_style(&selector));
+            }
+            EngineCmd::GetPage { reply } => {
+                let resp = eng.last_page.as_ref().map(|p| {
+                    page_to_api_response(p, &p.base_url.clone())
+                });
+                let _ = reply.send(resp);
+            }
+            EngineCmd::GetElements { reply } => {
+                let elems = eng.last_page.as_ref().map(|p| {
+                    page_to_api_response(p, &p.base_url.clone()).elements
+                }).unwrap_or_default();
+                let _ = reply.send(elems);
+            }
+            EngineCmd::LoadImage { url, bytes } => {
+                eng.image_cache.insert(url, bytes);
+            }
+            EngineCmd::Tick { timestamp, deadline, reply } => {
+                let needs = eng.tick_js(Some(timestamp), deadline);
+                let overrides = eng.get_style_overrides();
+                for (id, props) in overrides {
+                    eng.js_style_overrides.entry(id).or_default().extend(props);
+                }
+                let _ = reply.send(needs);
+            }
+            EngineCmd::Shutdown => break,
+        }
+    }
+}
+
+/// Cloneable handle used by GUI and HTTP threads to send commands to the engine actor.
+#[derive(Clone)]
+pub struct EngineHandle {
+    pub tx: mpsc::SyncSender<EngineCmd>,
+}
+
+impl EngineHandle {
+    /// Create a new engine actor and return a handle to it.
+    pub fn spawn() -> Self {
+        let (tx, rx) = mpsc::sync_channel::<EngineCmd>(64);
+        std::thread::Builder::new()
+            .name("engine-actor".into())
+            .spawn(move || run_engine_actor(rx))
+            .expect("failed to start engine actor thread");
+        Self { tx }
+    }
+
+    pub fn send_navigate(&self, url: String, width: f32) -> Result<PageResult, String> {
+        let (reply_tx, reply_rx) = mpsc::channel();
+        self.tx
+            .send(EngineCmd::Navigate { url, width, reply: reply_tx })
+            .map_err(|_| "engine disconnected".to_string())?;
+        reply_rx.recv().map_err(|_| "engine disconnected".to_string())?
+    }
+
+    pub fn send_re_render(
+        &self,
+        hovered_id: Option<String>,
+        focused_id: Option<String>,
+        width: f32,
+    ) -> Result<PageResult, String> {
+        let (reply_tx, reply_rx) = mpsc::channel();
+        self.tx
+            .send(EngineCmd::ReRender { hovered_id, focused_id, width, reply: reply_tx })
+            .map_err(|_| "engine disconnected".to_string())?;
+        reply_rx.recv().map_err(|_| "engine disconnected".to_string())?
+    }
+
+    pub fn send_get_page(&self) -> Option<ApiPageResponse> {
+        let (reply_tx, reply_rx) = mpsc::channel();
+        self.tx.send(EngineCmd::GetPage { reply: reply_tx }).ok()?;
+        reply_rx.recv().ok()?
+    }
+
+    pub fn send_get_elements(&self) -> Vec<ApiElement> {
+        let (reply_tx, reply_rx) = mpsc::channel();
+        if self.tx.send(EngineCmd::GetElements { reply: reply_tx }).is_err() {
+            return vec![];
+        }
+        reply_rx.recv().unwrap_or_default()
+    }
+
+    pub fn send_click(&self, x: f32, y: f32) -> Vec<ClickResult> {
+        let (reply_tx, reply_rx) = mpsc::channel();
+        if self.tx.send(EngineCmd::Click { x, y, reply: reply_tx }).is_err() {
+            return vec![];
+        }
+        reply_rx.recv().unwrap_or_default()
+    }
+
+    pub fn send_evaluate_js(&self, script: String) -> String {
+        let (reply_tx, reply_rx) = mpsc::channel();
+        if self.tx.send(EngineCmd::EvaluateJs { script, reply: reply_tx }).is_err() {
+            return String::new();
+        }
+        reply_rx.recv().unwrap_or_default()
+    }
+
+    pub fn send_screenshot(&self) -> Option<Vec<u8>> {
+        let (reply_tx, reply_rx) = mpsc::channel();
+        self.tx.send(EngineCmd::Screenshot { reply: reply_tx }).ok()?;
+        reply_rx.recv().ok()?
+    }
+
+    pub fn send_dom_tree(&self) -> String {
+        let (reply_tx, reply_rx) = mpsc::channel();
+        if self.tx.send(EngineCmd::DomTree { reply: reply_tx }).is_err() {
+            return String::new();
+        }
+        reply_rx.recv().unwrap_or_default()
+    }
+
+    pub fn send_layout_tree(&self) -> String {
+        let (reply_tx, reply_rx) = mpsc::channel();
+        if self.tx.send(EngineCmd::LayoutTree { reply: reply_tx }).is_err() {
+            return String::new();
+        }
+        reply_rx.recv().unwrap_or_default()
+    }
+
+    pub fn send_computed_style(&self, selector: String) -> HashMap<String, String> {
+        let (reply_tx, reply_rx) = mpsc::channel();
+        if self.tx
+            .send(EngineCmd::ComputedStyle { selector, reply: reply_tx })
+            .is_err()
+        {
+            return HashMap::new();
+        }
+        reply_rx.recv().unwrap_or_default()
+    }
+
+    pub fn send_tick(&self, timestamp: f64, deadline: Option<f64>) -> bool {
+        let (reply_tx, reply_rx) = mpsc::channel();
+        if self.tx.send(EngineCmd::Tick { timestamp, deadline, reply: reply_tx }).is_err() {
+            return false;
+        }
+        reply_rx.recv().unwrap_or(false)
+    }
+}
+
 // ── Tests ─────────────────────────────────────────────────────────────────────
 
 #[cfg(test)]
